+++ src/BasicInputs/index.ts
@@ -3,5 +3,6 @@ export * from './DatetimeInput';
 export * from './NumberInput';
 export * from './SelectInput';
 export * from './StringInput';
+export * from './CreatableSelectInput';
 
 export * from './ModelItemInput';